app/requirements.txt
alabaster==0.7.13
alembic==1.13.1
altgraph==0.17.3
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
Babel==2.12.1
blinker==1.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
debugpy==1.8.1
dnspython==2.7.0
docutils==0.18.1
dominate==2.9.1
email-validator==2.1.0.post1
Flask==3.0.3
Flask-Assets==2.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.4
flask-nav3==0.7.2
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
flask-sock==0.6.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
FormEncode==2.0.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
imagesize==1.4.1
importlib_resources==6.4.4
iniparse==0.5
itsdangerous==2.1.2
Jinja2==3.1.5
loutilities==3.9.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysqlclient==2.2.6
packaging==23.1
passlib==1.7.4
pefile==2023.2.7
ping3==4.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyserial==3.5
pyserial-asyncio==0.6
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.32.3
rich==13.9.4
rstcheck==6.2.4
rstcheck-core==1.2.1
setuptools==77.0.3
shellingham==1.5.3
simple-websocket==0.10.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinx_design==0.5.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.19
sqlalchemy-datatables==2.0.1
telnetlib3==2.0.4
text-unidecode==1.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.2.3
validators==0.35.0
visitor==0.1.3
webassets==2.0
websockets==11.0.3
Werkzeug==3.1.2
wsproto==1.2.0
WTForms==3.0.1
